Work on ECS system and renderer Deleting of entities, components, componentTable incl. flushing of temp data from a table Renderer now renders normals as color values

+++ GwaCore/src/ecs/TypeIDGenerator.h
@@ -3,15 +3,15 @@
 
 class TypeIDGenerator
 {
-    static size_t identifier() noexcept {
-        static size_t value = 0;
+    static uint32_t identifier() noexcept {
+        static uint32_t value = 0;
         return value++;
     }
 
 public:
     template<typename>
-    static size_t type() noexcept {
-        static const size_t value = identifier();
+    static uint32_t type() noexcept {
+        static const uint32_t value = identifier();
         return value;
     }
 };
